Refactored MainWindowViewModel
