+++ BitPoolMiner/ViewModels/WorkerViewModel.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BitPoolMiner.ViewModels
+{
+    class WorkerViewModel
+    {
+        private string workerName;
+        public string WorkerName { get; set; }
+    }
+}
